+++ src/keepTrackApi.ts
@@ -86,6 +86,7 @@ export class KeepTrackApi {
   on = EventBus.getInstance().on.bind(EventBus.getInstance());
   once = EventBus.getInstance().once.bind(EventBus.getInstance());
   emit = EventBus.getInstance().emit.bind(EventBus.getInstance());
+  emitAsync = EventBus.getInstance().emitAsync.bind(EventBus.getInstance());
   unregister = EventBus.getInstance().unregister.bind(EventBus.getInstance());
   events = EventBus.getInstance().events;
   methods = EventBus.getInstance().methods;